Extract TermsAndConditionsDialog component
Refactoring only!
