+++ src/web/components/TermsAndConditions/TermsAndConditions.scss
@@ -21,12 +21,6 @@
   }
 }
 
-.terms-conditions-dialog {
-  box-sizing: border-box;
-  height: 100%;
-  max-width: 910px;
-}
-
 .terms-and-conditions-form {
   flex-direction: column;
   align-items: center;
