+++ src/web/screens/dashboard.scss
@@ -9,8 +9,3 @@
   border-top: 1px solid #dcdee1;
   padding-top: 1.375rem;
 }
-
-.accept-error {
-  color: var(--theme-error);
-  margin-top: 12px;
-}
