+++ src/web/screens/dashboard.tsx
@@ -1,13 +1,10 @@
-import { useKeycloak } from '@react-keycloak/web';
-import { useContext, useEffect, useState } from 'react';
+import { useContext, useEffect } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
 
-import { Dialog } from '../components/Core/Dialog/Dialog';
 import { SideNav } from '../components/Navigation/SideNav';
-import { TermsAndConditionsForm } from '../components/TermsAndConditions/TermsAndConditions';
+import { TermsAndConditionsDialog } from '../components/TermsAndConditions/TermsAndConditionsDialog';
 import { CurrentUserContext } from '../contexts/CurrentUserProvider';
 import { ParticipantContext } from '../contexts/ParticipantProvider';
-import { SetTermsAccepted } from '../services/userAccount';
 import { RouteErrorBoundary } from '../utils/RouteErrorBoundary';
 import { ApiKeyManagementRoute } from './apiKeyManagement';
 import { ClientSideIntegrationRoute } from './clientSideIntegration';
@@ -43,21 +40,9 @@ const standardMenu = StandardRoutes.filter((r) => r.description);
 
 function Dashboard() {
   const { participant } = useContext(ParticipantContext);
-  const { LoggedInUser, loadUser } = useContext(CurrentUserContext);
-  const [showMustAccept, setShowMustAccept] = useState(false);
-  const { keycloak } = useKeycloak();
+  const { LoggedInUser } = useContext(CurrentUserContext);
   const navigate = useNavigate();
   const adminMenu = LoggedInUser?.user?.isApprover ? AdminRoutes.filter((r) => r.description) : [];
-
-  const handleAccept = async () => {
-    await SetTermsAccepted();
-    // Force token refresh after role updated
-    await keycloak.updateToken(10000);
-    await loadUser();
-  };
-  const handleCancel = () => {
-    setShowMustAccept(true);
-  };
 
   useEffect(() => {
     if (!participant) {
@@ -68,20 +53,7 @@ function Dashboard() {
     <div className='app-panel'>
       <SideNav standardMenu={standardMenu} adminMenu={adminMenu} />
       <div className='dashboard-content'>
-        {!LoggedInUser?.user?.acceptedTerms ? (
-          <Dialog className='terms-conditions-dialog'>
-            <TermsAndConditionsForm onAccept={handleAccept} onCancel={handleCancel}>
-              {showMustAccept && (
-                <div className='accept-error'>
-                  Please review the Terms and Conditions document. When you’ve scrolled to the
-                  bottom, click <b>Accept Terms and Conditions</b>.
-                </div>
-              )}
-            </TermsAndConditionsForm>
-          </Dialog>
-        ) : (
-          <Outlet />
-        )}
+        {!LoggedInUser?.user?.acceptedTerms ? <TermsAndConditionsDialog /> : <Outlet />}
       </div>
     </div>
   );
